afl-cc all-in-one
formated with custom-format.py
